+++ .gitmodules
@@ -1,3 +1,3 @@
 [submodule "pixelfonts"]
 	path = pixelfonts
-	url = git@github.com:sebseb7/pixelfonts.git
+	url = https://sebseb7@github.com/sebseb7/pixelfonts.git
