+++ libs/text.c
@@ -2,8 +2,8 @@
 #include <string.h>
 #include <stdio.h>
 
-#include "../pixelfonts/5x3/font.h"
-#include "../main.h"
+#include "pixelfonts/5x3/font.h"
+#include "main.h"
 #include "text.h"
 
 void print_5x3_at (int x,int y, char *text, char brightness)
