+++ r0ketr1s_sdl/sdl_draw
